fix(server): break the tts/index ↔ provider import cycle (partial-mock flake)
Root-causes the intermittent `[vitest] No "isTtsModelKey" export is defined on the "../tts/index.js" mock` that failed the v1.5.1 cross-OS release gate (Windows + macOS Verify).

`tts/index.ts` imported the GeminiTtsProvider / SidecarTtsProvider classes, while gemini.ts and sidecar.ts imported pure helpers (resolveGeminiModelId / sidecarModelId) back from `./index.js` — a runtime cycle. The 7 route tests that `vi.mock('../tts/index.js', async (importOriginal) => ({ ...await importOriginal(), ... }))` therefore resolved `index.js` mid-cycle, and under the parallel test:server pool (where qwen-voice.test.ts runs) the namespace came back partial, so the spread dropped an export.

Fix: move the pure, provider-independent declarations (TtsEngine, TtsModelKey, TTS_MODEL_LABELS, resolveGeminiModelId, isTtsModelKey, engineForModelKey, sidecarModelId) into a new leaf module `tts/model-keys.ts` (imports nothing). gemini.ts / sidecar.ts now import their helper from the leaf; index.ts re-exports it (`export * from './model-keys.js'`) so the public `tts/index.js` surface is unchanged for every importer. Runtime graph is now acyclic, so importOriginal('../tts/index.js') is always complete.

Behaviour-neutral. A new identity guard in index.test.ts asserts index re-exports the SAME helper references as the leaf, so a future edit can't move one back into index and re-form the cycle. The sibling toVoiceLike / synthesise-chapter.js log line is a different mechanism (no cycle; already single-forked on CI via test:server-slow) and is out of scope. Regression plan: docs/features/158-tts-index-cycle-mock-flake.md.
