+++ docs/features/INDEX.md
@@ -82,6 +82,7 @@ This section lists only the **in-flight working set** — plans whose `status:` 
 
 ### K. Cross-cutting invariants
 
+- [158 — Break the tts/index ↔ provider cycle (partial-mock release-gate flake)](158-tts-index-cycle-mock-flake.md) — `stable`. Root-causes the intermittent `[vitest] No "isTtsModelKey" export is defined on the "../tts/index.js" mock` that failed the v1.5.1 cross-OS gate (win+mac). `tts/index.ts` imported the provider classes while `gemini.ts`/`sidecar.ts` imported pure helpers back from `index.js` — a runtime cycle, so `importOriginal('../tts/index.js')` in the 7 route tests that mock it returned a partial namespace under the parallel pool. Fix: move the pure helpers (`isTtsModelKey`, `engineForModelKey`, `sidecarModelId`, `resolveGeminiModelId`, `TTS_MODEL_LABELS`, `TtsModelKey`/`TtsEngine`) to a new leaf `tts/model-keys.ts`; providers import the leaf; `index.ts` re-exports it (`export *`) so the public surface is unchanged. Identity guard in `index.test.ts` locks the re-export so the cycle can't re-form.
 - [157 — Default-bind the HTTP server to loopback (srv-19)](157-srv19-loopback-bind.md) — `draft`. Closes 2026-05-31 security-review findings #1+#2: plain-HTTP dev mode bound `0.0.0.0`, so every unauthenticated route + the `/workspace` static mount (all manuscripts/audio/`state.json`/`cast.json`) was reachable by any LAN peer. New pure `selectBindHost(lanHttps, env)` helper (`server/src/bind-host.ts`) → default bind `127.0.0.1`; the opt-in `LAN_HTTPS` mobile flow keeps binding all interfaces (deliberately reachable); `BIND_HOST`/`HOST` env override restores all-interface HTTP for power users. One host arg threaded through the existing `index.ts` listen wiring; `export-lan` LAN/port/CORS tests stay green. Cheapest hardening from the review; the rest (srv-20/21/22, side-12, ops-7) deferred.
 - [156 — Pre-commit scope filter + GPU-contention test throttle](156-precommit-scope-contention.md) — `active`. Pre-commit (`verify:fast:scoped`) now skips any leg whose scope the **staged diff** never touched (`scripts/verify-cache.mjs --scope-staged`, mirroring `verify.yml`'s scope matcher), so a sidecar-only/docs-only commit runs none of the frontend/server suites — closing the green-only-cache hole where a contention flake forced out-of-scope re-runs. Plus a soft GPU-contention guard: `verify-cache.mjs` probes `nvidia-smi` and sets `LOW_CONCURRENCY=1` (frontend pool → half-cores, server `maxForks` 2→1; otherwise plan-45 defaults) when a generation run is hammering the box. Pre-push still runs the full battery.
 - [118 — CI cost reduction round 2 (run-count + test-impact)](118-ci-cost-round-2.md) — Stacks on plans 103/106. Four levers against Linux runner-minute spend: (1) explicit `timeout-minutes` on every job (`verify` 15→20, `pr-title-lint` 5, `release` verify 20 / publish 15) to kill the timeout double-charge + 360-min runaway; (2) **draft-by-default** convention — `verify.yml` already skips drafts + re-fires on `ready_for_review`, so develop-on-draft → local verify → `gh pr ready` bills ONE run per PR; (3) **one integration PR per parallel-agent round**, verified once, instead of N PRs; (4) **test-impact selection** — the frontend + server vitest legs run `vitest --changed <base>` (diff-affected tests only), with `forceRerunTriggers` on the runner-injected `src/test/setup.ts` closing the only module-graph gap. e2e stays scope-gated (Playwright `--only-changed` is unsafe for app-source changes). Full suite preserved on local pre-push + weekly cross-os.
